Move function `findClosetValue`
