+++ src/lib/components/molecules/content-table/content-table.tsx
@@ -4,9 +4,9 @@ import type { Component } from "#/lib/utils/component";
 import type { ContentTableProps } from "./content-table.type";
 import { useEffect, useRef, useState } from "react";
 import clsx from "clsx";
-import { findClosestValue } from "#/lib/utils/array";
 import { IoIosArrowUp } from "react-icons/io";
 import { ContentTableItem } from "./content-table-item";
+import { findClosestValue } from "./content-table.util";
 
 export const ContentTable: Component<ContentTableProps> = ({ items }) => {
   const [isCtOpen, setIsCtOpen] = useState(false);
@@ -25,10 +25,7 @@ export const ContentTable: Component<ContentTableProps> = ({ items }) => {
 
   const scrollToTitle = (title: string): void => {
     const elements = Array.from(document.querySelectorAll<HTMLElement>("h1, h2"));
-    const match = elements.find((el) => {
-      return  el.textContent?.toLowerCase().includes(title.toLowerCase());
-    });
-
+    const match = elements.find((el) => el.textContent?.toLowerCase().includes(title.toLowerCase()));
     if (!match) return;
 
     window.scroll(0, match.offsetTop - 100);
@@ -36,9 +33,7 @@ export const ContentTable: Component<ContentTableProps> = ({ items }) => {
 
   useEffect(() => {
     const elements = Array.from(document.querySelectorAll<HTMLElement>("h1, h2"));
-    const usedElements = elements.filter((el) => {
-      return items.some((itemEl) => itemEl.toLowerCase() === el.textContent?.toLowerCase());
-    });
+    const usedElements = elements.filter((el) => items.some((itemEl) => itemEl.toLowerCase() === el.textContent?.toLowerCase()));
 
     const onScroll = (): void => {
       const closestElement = findClosestValue(usedElements, (window.scrollY + 115));
@@ -62,27 +57,26 @@ export const ContentTable: Component<ContentTableProps> = ({ items }) => {
       <div className="flex items-center justify-between">
         <p className="text-white font-medium text-lg">Table des matières</p>
         <IoIosArrowUp className={clsx(
-          "text-white h-5 w-5 lg:hidden transition-transform",
-          {
+          "text-white h-5 w-5 lg:hidden transition-transform", {
             "rotate-180": isCtOpen
           }
         )} onClick={() => setIsCtOpen((state) => !state)} />
       </div>
       <div className={clsx(
-        "lg:max-h-96 overflow-hidden transition-[max-height]",
-        {
+        "lg:max-h-96 overflow-hidden transition-[max-height]", {
           "max-h-0": !isCtOpen,
           "max-h-96": isCtOpen
         }
       )}>
         <span className={clsx(
-          "absolute bg-purple rounded-full w-1 h-4 mt-3 transition-transform left-4 lg:opacity-100 transition-opacity",
-          {
+          "absolute bg-purple rounded-full w-1 h-4 mt-3 transition-transform left-4 lg:opacity-100 transition-opacity", {
             "opacity-0": !isCtOpen,
             "opacity-100": isCtOpen
           }
         )} ref={linkPointerRef} />
-        <div className="grid gap-1 pt-2" ref={itemsContainerRef}>
+        <div
+          aria-label="content-table"
+          className="grid gap-1 pt-2" ref={itemsContainerRef}>
           {items.map((item, i) => (
             <ContentTableItem
               key={i}
